Added components folder and moved search into it
