+++ world-bank-app/src/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import Search from "./Search";
+import Search from "./components/Search";
 import "bootstrap/dist/css/bootstrap.min.css";
 
 function App() {
